feat: Use config from cnd, rename default config
