+++ src/javascript/CKEditor/configurations/index.js
@@ -1,4 +1,4 @@
-export * from './config-default';
+export * from './config-complete';
 export * from './config-minimal';
 export * from './config-light';
 export * from './config-advanced';
