+++ src/javascript/CKEditor/configurations/plugins-complete.js
